requirements.txt


Flask==2.3.3
Flask-Login==0.6.3
psycopg2==2.9.9  
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0